refactor: TripApi 관련 Dto -> TripApiDataDto로 변경 후 패키지 이동 일회성으로 사용하므로 response, request 나누지 않고 DTO  하나로 사용
